requirements.txt
aniso8601==1.1.0
beautifulsoup4==4.3.2
cssselect==0.9.1
feedfinder2==0.0.1
feedparser==5.1.3
Flask==0.10.1
Flask-RESTful==0.3.5
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.8
lxml==3.3.5
MarkupSafe==0.23
newspaper3k==0.1.5
nltk==3.0.1
Pillow==2.6.1
python-dateutil==2.4.0
pytz==2015.7
PyYAML==3.11
requests==2.3.0
six==1.7.3
tldextract==1.5.1
Werkzeug==0.11.3
wheel==0.24.0
